Move ContentTypeNegotiation to separate file
