+++ core/src/main/scala/io/finch/ToResponse.scala
@@ -3,7 +3,6 @@ package io.finch
 import com.twitter.concurrent.AsyncStream
 import com.twitter.finagle.http.{Response, Status, Version}
 import com.twitter.io.Buf
-import io.finch.internal.Accept
 import java.nio.charset.Charset
 import scala.annotation.implicitNotFound
 import shapeless._
@@ -15,50 +14,6 @@ trait ToResponse[A] {
   type ContentType
 
   def apply(a: A, cs: Charset): Response
-}
-
-trait ContentTypeNegotiation[ContentType] {
-
-  type A
-
-  def apply(accept: Seq[Accept]): ToResponse.Aux[A, ContentType]
-
-}
-
-object ContentTypeNegotiation extends LowPriorityNegotiation {
-
-  type Aux[CT, α] = ContentTypeNegotiation[CT] { type A = α }
-
-  implicit def mkCoproduct[A, CTH <: String, CTT <: Coproduct](implicit
-   h: ToResponse.Aux[A, CTH],
-   t: ContentTypeNegotiation.Aux[CTT, A],
-   w: Witness.Aux[CTH]
-  ): ContentTypeNegotiation.Aux[CTH :+: CTT, A] = instance { accept =>
-    Accept(w.value) match {
-      case Some(ct) if accept.exists(_.matches(ct)) =>
-        ToResponse.instance[A, CTH :+: CTT]((a, cs) => h(a, cs))
-      case _ =>
-        ToResponse.instance[A, CTH :+: CTT]((a, cs) => t(accept)(a, cs))
-    }
-  }
-
-}
-
-trait LowPriorityNegotiation {
-
-  def instance[CT, α](f: Seq[Accept] => ToResponse.Aux[α, CT]): ContentTypeNegotiation.Aux[CT, α] = {
-    new ContentTypeNegotiation[CT] {
-      type A = α
-      def apply(accept: Seq[Accept]): ToResponse.Aux[A, CT] = f(accept)
-    }
-  }
-
-  implicit def mkLast[A, CTH <: String](implicit
-   h: ToResponse.Aux[A, CTH]
-  ): ContentTypeNegotiation.Aux[CTH :+: CNil, A] = instance { _ =>
-    ToResponse.instance[A, CTH :+: CNil]((a, cs) => h(a, cs))
-  }
-
 }
 
 trait LowPriorityToResponseInstances {
@@ -100,19 +55,19 @@ trait LowPriorityToResponseInstances {
   implicit def jsonAsyncStreamToResponse[A](implicit
     e: Encode.Json[A]
   ): Aux[AsyncStream[A], Application.Json] =
-    asyncResponseBuilder((a, cs) => e(a, cs).concat(NewLine))
+    asyncResponseBuilder[A, Application.Json]((a, cs) => e(a, cs).concat(NewLine))
 
   implicit def textAsyncStreamToResponse[A](implicit
     e: Encode.Text[A]
   ): Aux[AsyncStream[A], Text.Plain] =
-    asyncResponseBuilder((a, cs) => e(a, cs).concat(NewLine))
+    asyncResponseBuilder[A, Text.Plain]((a, cs) => e(a, cs).concat(NewLine))
 }
 
 trait HighPriorityToResponseInstances extends LowPriorityToResponseInstances {
 
   implicit def asyncBufToResponse[CT <: String](implicit
     w: Witness.Aux[CT]
-  ): Aux[AsyncStream[Buf], CT] = asyncResponseBuilder((a, _) => a)
+  ): Aux[AsyncStream[Buf], CT] = asyncResponseBuilder[Buf, CT]((a, _) => a)
 
   implicit def responseToResponse[CT <: String]: Aux[Response, CT] = instance((r, _) => r)
